requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
xgboost==1.7.6
torch==2.0.1
mlflow==2.5.0
fastapi==0.100.0
uvicorn==0.22.0
optuna==3.2.0
prefect==2.10.0
pyyaml==6.0
joblib==1.3.1
